+++ test/Main.hs
@@ -1,10 +1,12 @@
 module Main (main) where
 
 import Control.Monad (unless, when)
+import Data.ByteString (ByteString)
 import Data.Char (ord)
 import Data.Kind (Type)
 import Data.Text (Text)
 import Data.Text qualified as Text
+import GHC.Exts (fromList)
 import GHC.IO.Encoding (setLocaleEncoding, utf8)
 import Plutarch (
   Config (NoTracing, Tracing),
@@ -21,6 +23,7 @@ import Plutarch.Prelude (
   pconstant,
   plam,
   plift,
+  pnot,
   ptraceDebug,
   ptraceDebugError,
   ptraceDebugIfFalse,
@@ -32,9 +35,16 @@ import Plutarch.Prelude (
   ptraceInfoIfTrue,
   ptraceInfoShowId,
   (#),
+  (#$),
   (#==),
  )
-import Test.Tasty (defaultMain, testGroup)
+import Plutarch.String (pisHexDigit)
+import Test.Tasty (
+  DependencyType (AllSucceed),
+  after,
+  defaultMain,
+  testGroup,
+ )
 import Test.Tasty.HUnit (assertEqual, assertFailure, testCase)
 
 main :: IO ()
@@ -49,6 +59,16 @@ main = do
                 (pallBS # plam (#== pconstant (toInteger . ord $ 'a')) # pconstant "aaaaaaaaaa")
             , testCase "predicate missing one case fails" . isFalse $
                 (pallBS # plam (#== pconstant (toInteger . ord $ 'a')) # pconstant "aaaaaaaaab")
+            ]
+        , after AllSucceed "pallBS" . testGroup "pisHexDigit" $
+            [ testCase "numbers are hex digits" . isTrue $
+                (pallBS # pisHexDigit # pconstant "0123456789")
+            , testCase "A-F are hex digits" . isTrue $
+                (pallBS # pisHexDigit # pconstant "ABCDEF")
+            , testCase "a-f are hex digits" . isTrue $
+                (pallBS # pisHexDigit # pconstant "abcdef")
+            , testCase "no other ASCII code is a hex digit" . isTrue $
+                (pallBS # plam (\x -> pnot #$ pisHexDigit # x) # pconstant nonHexAscii)
             ]
         ]
     , testGroup
@@ -169,3 +189,17 @@ isFalse comp = case evalTerm NoTracing comp of
   Right (res, _, _) -> case res of
     Left err -> assertFailure $ "Execution errored: " <> show err
     Right t -> when (plift t) (assertFailure "is true")
+
+nonHexAscii :: ByteString
+nonHexAscii =
+  -- All codes up to, but not including, the first digit
+  fromList [0, 1 .. 47]
+    <>
+    -- Between digits to upper-case
+    fromList [58, 59 .. 64]
+    <>
+    -- Between upper-case and lower-case
+    fromList [71, 72 .. 96]
+    <>
+    -- After lower-case
+    fromList [103 .. 127]