+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## Added
 
-* `pallBS` to `Plutarch.ByteString`
+* `pallBS` to `Plutarch.ByteString` (originally from `plutarch-extra')
+* `pisHexDigit` to `Plutarch.String` (originally from `plutarch-extra`)
 
 # 1.9.0 - 25-09-2024